consolidate report_ctd into ctd_io
